--- practice_2.py
@@ -1,12 +1,0 @@
-class clacla:
-    def __init__(self):
-        self.name = "namename"
-        self.age = "ageage"
-
-    def pripri(self):
-        print(self.name, self.age)
-
-
-class ovov:
-    def overover(self):
-        print("before over")
